Move modal components into their own dir
