+++ src/routes/details/components/cluster/modal/clusterContent.tsx
