+++ src/routes/details/components/cluster/modal/clusterModal.scss
