+++ src/routes/details/components/cluster/modal/clusterModal.styles.ts
