+++ src/routes/details/components/cluster/modal/clusterModal.tsx
