+++ src/routes/details/components/summary/modal/summaryModal.scss
