+++ src/routes/details/components/summary/modal/summaryModal.styles.ts
