+++ src/routes/details/components/tag/tag.styles.ts
@@ -1,16 +1,8 @@
-import global_FontWeight_bold from '@patternfly/react-tokens/dist/js/global_FontWeight_bold';
 import global_spacer_3xl from '@patternfly/react-tokens/dist/js/global_spacer_3xl';
-import global_spacer_lg from '@patternfly/react-tokens/dist/js/global_spacer_lg';
 import global_spacer_sm from '@patternfly/react-tokens/dist/js/global_spacer_sm';
 import type React from 'react';
 
 export const styles = {
-  dataListHeading: {
-    fontWeight: global_FontWeight_bold.value as any,
-  },
-  groupByHeading: {
-    marginBottom: global_spacer_lg.value,
-  },
   tagLink: {
     marginLeft: global_spacer_sm.value,
   },
